HACKY FIX: Use stock V30 msm8998 audio primary
This is being done as our msm8996 audio primary does not contain the required functions needed fo adjusting filters on the dac.
So, I have taken the stock audio primary from the V30 to allow for controlling the filters on the dac. The msm8996 and msm8998 both use the WCD9335 audio codec so there should be no issues. V30 and V20 also use an almost identical dac, the ES9218 on the V20 and ES9218P on the V30.

The audio.primary.msm8998 lib has been renamed to audio.primary.msm8996.